refactor: Comprehensive example-cli refactoring for maintainability
This commit implements a systematic refactoring of example-cli to improve
code organization, reduce duplication, and enhance maintainability before
Metal GPU integration.

## Changes

### 1. Configuration Merge Refactoring
- Add `GenerationConfig::merge_from_cli()` method in session/config.rs
- Eliminate 30 lines of duplicated merge logic from main.rs
- Add serde default attributes and helper functions
- Fix type consistency (u32 for top_k across codebase)
- Update test expectations to match correct defaults (2048 tokens)

### 2. Backend Loader Separation (Primary Achievement)
- Create new `model/backend_loader.rs` module (230 lines)
- Implement `Architecture` enum for GPT/Llama detection
- Implement `BackendLoader` with unified `load()` interface
- Extract 232 lines of duplicated backend loading from main.rs
- Reduce main.rs from 471 → 239 lines (49% reduction)
- Support hybrid-f32, mac-hybrid, and standard backends

### 3. InferenceEngine Simplification
- Introduce `ModelBackend` enum to unify model storage
- Replace 4 Option<Model> fields with single Option<ModelBackend>
- Simplify model checking logic (remove complex conditionals)
- Refactor generate_tokens() with clean match patterns
- Update F32 model methods to use ModelBackend enum
- Improve code clarity and type safety

### 4. Documentation
- Add CLI_REFACTORING_PLAN.md with detailed refactoring roadmap
- Add METAL_INTEGRATION_STATUS.md documenting Metal implementation
- Add BACKEND_INTEGRATION_PLAN.md for future backend work

## Impact

**Code Reduction:**
- main.rs: 471 → 239 lines (-232 lines, -49%)
- Total reduction: 262 lines of duplicated code eliminated

**Code Addition:**
- backend_loader.rs: +230 lines (new module)
- Documentation: +980 lines

**Improvements:**
- DRY principle applied (no code duplication)
- Separation of concerns (backend loading isolated)
- Enhanced maintainability (new backends easier to add)
- Improved type safety (ModelBackend enum)
- Better code organization and readability

## Testing
- ✅ Build successful (cargo build --release --features metal)
- ✅ CLI functionality verified
- ✅ No breaking changes to external API

+++ example-cli/src/cli/repl.rs
@@ -427,7 +427,7 @@ impl REPL {
         config.generation.max_tokens = gen_config.max_tokens;
         config.generation.temperature = gen_config.temperature;
         config.generation.top_p = gen_config.top_p;
-        config.generation.top_k = gen_config.top_k as usize;
+        config.generation.top_k = gen_config.top_k;
 
         // Set backend
         config.backend.default = self.session.backend_name().to_string();
